TVShowDetailViewModelTest

+++ app/src/test/java/com/arifikhsan/jetpackmoviecatalogue/ui/movies/detail/MovieDetailViewModelTest.kt
@@ -63,6 +63,6 @@ class MovieDetailViewModelTest {
         assertEquals(sampleMovie.voteCount, movie?.voteCount)
 
         assertEquals(sampleMovie.popularity as Double, movie?.popularity as Double, 0.0001)
-        assertEquals(sampleMovie.voteAverage as Double, movie.voteAverage as Double, 0.0001)
+        assertEquals(sampleMovie.voteAverage as Double, movie.voteAverage, 0.0001)
     }
 }
